Added phone input component to update
