+++ src/components/forms/update-identity-form.tsx
@@ -4,6 +4,7 @@ import React from "react";
 import styled from "styled-components";
 import { v4 as uuidv4 } from "uuid";
 
+import { StyledPhoneInput } from "../input/styled-phone-input";
 import { WrongInputError } from "../input/wrong-input-error";
 import { Form } from "./form";
 import { HttpVerbs } from "@src/@types/http-verbs";
@@ -14,6 +15,7 @@ import { FakeButton } from "@src/components/buttons/fake-button";
 import { Input } from "@src/components/input/input";
 import { NeutralLink } from "@src/components/neutral-link/neutral-link";
 import { fetchJson } from "@src/utils/fetch-json";
+import { getIdentityType } from "@src/utils/get-identity-type";
 
 const UpdateIdentityForm: React.FC<{
   currentIdentity: Identity;
@@ -68,24 +70,40 @@ const UpdateIdentityForm: React.FC<{
       >
         <p>
           New{" "}
-          {currentIdentity.type === IdentityTypes.PHONE
+          {getIdentityType(currentIdentity.type) === IdentityTypes.PHONE
             ? "phone number *"
             : "email address *"}
         </p>
-        <Input
-          type="text"
-          name="value"
-          placeholder={`Enter your ${currentIdentity.type}`}
-          value={identity.value}
-          onChange={(event) =>
-            setIdentity({
-              value: event.target.value,
-              type: currentIdentity.type,
-              expiresAt: Date.now() + 300000,
-              primary: currentIdentity.primary,
-            })
-          }
-        />
+        {getIdentityType(currentIdentity.type) === IdentityTypes.EMAIL ? (
+          <Input
+            type="text"
+            name="value"
+            placeholder={`Enter your ${currentIdentity.type}`}
+            value={identity.value}
+            onChange={(event) =>
+              setIdentity({
+                value: event.target.value,
+                type: currentIdentity.type,
+                expiresAt: Date.now() + 300000,
+                primary: currentIdentity.primary,
+              })
+            }
+          />
+        ) : (
+          <StyledPhoneInput
+            placeholder="Enter your phone number"
+            value={identity.value}
+            defaultCountry="FR"
+            onChange={(value) => {
+              setIdentity({
+                value,
+                type: currentIdentity.type,
+                expiresAt: Date.now() + 300000,
+                primary: identity.primary,
+              });
+            }}
+          />
+        )}
         <Button variant={ButtonVariant.PRIMARY} type="submit">
           Update {currentIdentity.type}
         </Button>
